+++ src/bound.rs
@@ -1,6 +1,6 @@
 use crate::r#type::Type;
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Eq, PartialEq)]
 pub struct Bound {
     pub name: String,
     pub bound: Vec<Type>,